+++ bob/src/link_manager.rs
@@ -1,3 +1,8 @@
+use tokio::sync::{
+    mpsc::{channel, Receiver, Sender},
+    RwLock,
+};
+
 use crate::prelude::*;
 
 use termion::color;
@@ -6,6 +11,7 @@ use termion::color;
 pub(crate) struct LinkManager {
     nodes: Arc<[Node]>,
     check_interval: Duration,
+    node_check_queue: Arc<RwLock<Option<Sender<String>>>>,
 }
 
 pub(crate) type ClusterCallOutput<T> = Result<NodeOutput<T>, NodeOutput<Error>>;
@@ -17,6 +23,7 @@ impl LinkManager {
         LinkManager {
             nodes: Arc::from(nodes),
             check_interval,
+            node_check_queue: Arc::new(RwLock::new(None)),
         }
     }
 
@@ -46,9 +53,40 @@ impl LinkManager {
         }
     }
 
-    pub(crate) fn spawn_checker(&self, factory: Factory) {
-        let nodes = self.nodes.clone();
-        tokio::spawn(Self::checker_task(factory, nodes, self.check_interval));
+    async fn priority_nodes_checker(
+        nodes: Arc<[Node]>,
+        factory: Factory,
+        mut node_check_queue: Receiver<String>,
+    ) {
+        while let Some(name) = node_check_queue.recv().await {
+            if let Some(node) = nodes.iter().find(|n| n.name() == name) {
+                if let Err(e) = node.check(&factory).await {
+                    error!(
+                        "No connection to {}:[{}] - {}",
+                        node.name(),
+                        node.address(),
+                        e
+                    );
+                } else {
+                    debug!("Create connection in response to ping from {}", node.name());
+                }
+            }
+        }
+    }
+
+    pub(crate) async fn spawn_checker(&self, factory: Factory) {
+        let (sender, receiver) = channel(self.nodes.len() * 10);
+        self.node_check_queue.write().await.replace(sender);
+        tokio::spawn(Self::checker_task(
+            factory.clone(),
+            self.nodes.clone(),
+            self.check_interval,
+        ));
+        tokio::spawn(Self::priority_nodes_checker(
+            self.nodes.clone(),
+            factory,
+            receiver,
+        ));
     }
 
     pub(crate) async fn call_nodes<F, T>(
@@ -86,5 +124,17 @@ impl LinkManager {
             Box::pin(client.exist(keys.to_vec(), GetOptions::new_all()))
         })
         .await
+    }
+
+    pub(crate) async fn update_node_connection(&self, node_name: &str) {
+        if let Some(node) = self.nodes.iter().find(|n| n.name() == node_name) {
+            if let Some(queue) = self.node_check_queue.read().await.as_ref() {
+                if !node.connection_is_set() {
+                    if let Err(e) = queue.send(node_name.to_string()).await {
+                        warn!("error while updating node {} status: {}", node_name, e);
+                    }
+                }
+            }
+        }
     }
 }