+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Bob versions changelog
 - Include bobc into release builds (#569)
 - Added blob-info and index-info features to brt (#356)
 - Support for files, file name patterns, key ranges and a 'exists' subcommand to 'bobc' (#539)
+- Add ping response for faster nodes reconnection (#625)
 
 #### Changed
 - Change locks to sync where possible (#472)